requirements.txt
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.1
faiss-cpu>=1.7.4
pypdf>=3.0.0
python-dotenv>=1.0.0
protobuf==3.20.3
sentence-transformers==2.2.2
transformers==4.37.2
numpy<2.0
langchain-huggingface>=0.0.1



